refactor: 캘린더 컴포넌트 위치 변경 및 TODO 적어두기
