+++ src/components/MissionDetail/MissionHistoryTab.tsx
@@ -1,7 +1,7 @@
 import { useParams } from 'next/navigation';
-import MissionCalendar from '@/app/mission/[id]/detail/MissionCalender/MissionCalendar';
 import MissionHistoryBannerApi from '@/app/mission/[id]/detail/MissionHistoryBanner/MissionHistoryBannerApi';
-import { css } from '@styled-system/css';
+import MissionCalendar from '@/components/MissionDetail/MissionCalender/MissionCalendar';
+import MissionHistoryTabLayout from '@/components/MissionDetail/MissionHistoryTabLayout';
 import dayjs from 'dayjs';
 
 function MissionHistoryTab({ isFollow }: { isFollow?: boolean }) {
@@ -10,28 +10,11 @@ function MissionHistoryTab({ isFollow }: { isFollow?: boolean }) {
   const currentData = dayjs();
 
   return (
-    <div className={scrollAreaCss}>
-      <div className={missionHistoryTabCss}>
-        {missionId && <MissionHistoryBannerApi missionId={missionId} />}
-        <MissionCalendar isFollow={isFollow} currentData={currentData} missionId={Number(missionId)} />
-      </div>
-    </div>
+    <MissionHistoryTabLayout>
+      {missionId && <MissionHistoryBannerApi missionId={missionId} />}
+      <MissionCalendar isFollow={isFollow} currentData={currentData} missionId={Number(missionId)} />
+    </MissionHistoryTabLayout>
   );
 }
 
 export default MissionHistoryTab;
-
-const scrollAreaCss = css({
-  overflowY: 'scroll',
-  height: 'calc(100vh - 24px - 44px)',
-  _scrollbar: {
-    display: 'none',
-  },
-});
-
-const missionHistoryTabCss = css({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '20px',
-  padding: '24px 16px 192px 16px',
-});
